more

+++ Python Certification Practice/Question 39.py
@@ -1,8 +1,8 @@
-distance = float(input("Enter the distance traveled in feet: "))
+distance = int(input("Enter the distance traveled in feet: "))
 
 distance_miles = distance/ 5280    #convert to miles
 
-time = int(input("Enter the time elapsed in seconds: "))
+time = float(input("Enter the time elapsed in seconds: "))
 
 time_hours = time/3600     #convert to hours
